fix: refacto packages/components/main/Home.tsx
