+++ js/packages/components/main/home/Footer.tsx
